docker, postgres, routes

+++ src/artists/artist.controller.ts
@@ -8,46 +8,72 @@ import {
   Body,
   HttpCode,
   HttpStatus,
+  BadRequestException,
+  NotFoundException,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
 import { ArtistService } from './artist.service';
-import { Artist } from './artist.interface';
 import { CreateArtistDto } from './create-artist.dto';
 import { UpdateArtistDto } from './update-artist.dto';
+import { Artist } from './artist.interface';
+import { validate as uuidValidate } from 'uuid';
 
 @Controller('artist')
 export class ArtistController {
   constructor(private readonly artistService: ArtistService) {}
 
   @Get()
-  getAll(): Artist[] {
+  async getAll(): Promise<Artist[]> {
     return this.artistService.findAll();
   }
 
   @Get(':id')
-  getById(@Param('id') id: string): Artist {
-    return this.artistService.findById(id);
+  async getOne(@Param('id') id: string): Promise<Artist> {
+    if (!uuidValidate(id)) {
+      throw new BadRequestException('Invalid UUID');
+    }
+    const artist = await this.artistService.findOne(id);
+    if (!artist) {
+      throw new NotFoundException('Artist not found');
+    }
+    return artist;
   }
 
   @Post()
   @UsePipes(new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true }))
-  create(@Body() createArtistDto: CreateArtistDto): Artist {
-    return this.artistService.create(createArtistDto);
+  @HttpCode(HttpStatus.CREATED)
+  async create(@Body() createArtistDto: CreateArtistDto) {
+    const artist = await this.artistService.create(createArtistDto);
+    return {
+      id: artist.id,
+      name: artist.name,
+      grammy: artist.grammy,
+    };
   }
 
   @Put(':id')
   @UsePipes(new ValidationPipe({ whitelist: true, forbidNonWhitelisted: true }))
-  update(
+  async update(
     @Param('id') id: string,
     @Body() updateArtistDto: UpdateArtistDto,
-  ): Artist {
-    return this.artistService.update(id, updateArtistDto);
+  ): Promise<Artist> {
+    if (!uuidValidate(id)) {
+      throw new BadRequestException('Invalid UUID');
+    }
+    const updatedArtist = await this.artistService.update(id, updateArtistDto);
+    if (!updatedArtist) {
+      throw new NotFoundException('Artist not found');
+    }
+    return updatedArtist;
   }
 
   @Delete(':id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  delete(@Param('id') id: string): void {
-    this.artistService.delete(id);
+  async delete(@Param('id') id: string): Promise<void> {
+    if (!uuidValidate(id)) {
+      throw new BadRequestException('Invalid UUID');
+    }
+    await this.artistService.delete(id);
   }
 }
